Reworked map to map area
The map has now an custom view the map area which handles and controls the views. The map area can be added to a view and allows markers to be added to it. For this the markers were also rewritten to use map positions, all this added some additional methods to use. Markers now also properly use the constructor method instead of create.