Reworked marker data management
Removed load method and location passer of game save and alike.
Added dynamic location interface to check if an activity allows location observers to be added.
Reworked save dialogs to be simplified and be the minimum what they need to be to work.
Update marker instantiating, a marker can now be loaded on a map area. This creates and add the image view of the marker.