Improved selected and navigation bar
Navigation bar is now not tied to the activity, only requirement is that activity includes navigation bar layout and configures the navigation bar with its activity and page type.
Selection now does not use intent to pass on the selected, which would copy it. but can request the activity such as edit activity for the selected.