Update navigation
Update navigation, activities are now finished except those that implement base activity interface of navigate. Additionally you can navigate to the same activity, which recreates it. (Useful for inspect and edit activity)
Clicking on images now takes you to the relevant pages, so clicking image in preview or in hint is navigable.
Fixed mistake in fruit, did not add empty array list yet to get content.