Update save methods
Methods annotated by Worker thread should now be actually called from a thread, so they don't hold back the main thread. Important to note is that one should be careful assuming the perceived state of the save platform is as what will be. Use the on finish runnable for updating the UI after the method is finished.
Added util for running on UI thread without context, also relocated the Util class.