Update edit activity
Content can now be removed on the edit activity form the selectable.
For this content remove was added which adds a button remove to the side of the content.
The button remove and add were updated to be more easily reused.
Added set content method to container interface, which allows the edit to update the new content list.