Added create (content) dialog
Allow content to be created/added on the edit activity. For this a more generic create dialog was made, which could also be used for the create fruit dialog and more in the future.

For this containers now have the remove and add content method.
Added content type, which determines what content can be created in the create dialog.
Update the edit activity to use the create content dialog.
Update edit hint to create a new hint and add it to the container.
Added layout and images, for create dialog and the content types.