Reworked map and more
The map has not a map storage and a map save which are the save objects that can be serialized and are disconnected from context.
The map markers has been reworked to be a part of the map save and thus also disconnected from context.
Markers now have a marker view which is the instantiation of the marker for the context and map area, which disconnects the data class from the context.
Confirm and create dialog are added as base classes to implement for instantiations of them to reduce duplication in the common dialog type.
From this big rework there are new holes to be filled to make things easier, namely methods in the map save and map area to get access and easy control over related elements.
Furthermore the implementation of the reworked elements need to be streamlined more, but at least there is not the opportunity to do so.